Update exam and exam result models to use DecimalField for marks; enh…

+++ backend/teacher/models.py
@@ -65,8 +65,8 @@ class Exam(models.Model):
         'classes.SchoolSection', on_delete=models.CASCADE
     )
     subject = models.ForeignKey(Subject, on_delete=models.CASCADE)
-    max_marks = models.PositiveIntegerField()
-    pass_marks = models.PositiveIntegerField()
+    max_marks = models.DecimalField(max_digits=20, decimal_places=10)
+    pass_marks = models.DecimalField(max_digits=20, decimal_places=10)
     exam_date = models.DateField(null=True)
     is_active = models.BooleanField(default=True)
     created_by_teacher = models.ForeignKey(Teacher, on_delete=models.CASCADE, related_name='exams_created')
@@ -83,7 +83,7 @@ class Exam(models.Model):
 class ExamResult(models.Model):
     exam = models.ForeignKey(Exam, on_delete=models.CASCADE)
     student = models.ForeignKey('student.Student', on_delete=models.CASCADE)
-    marks_obtained = models.DecimalField(max_digits=6, decimal_places=2)
+    marks_obtained = models.DecimalField(max_digits=20, decimal_places=10)
     created_by_teacher = models.ForeignKey(Teacher, on_delete=models.CASCADE, related_name='exam_results_created')
     updated_by_teacher = models.ForeignKey(Teacher, on_delete=models.CASCADE, null=True, blank=True, related_name='exam_results_updated')
     created_at = models.DateTimeField(auto_now_add=True)
